+++ test/fixtures/database.sql
@@ -2,3 +2,4 @@ DROP DATABASE IF EXISTS postgrest_test;
 DROP ROLE IF EXISTS postgrest_test;
 CREATE USER postgrest_test createdb createrole;
 CREATE DATABASE postgrest_test OWNER postgrest_test;
+alter database postgrest_test set postgrest.claims.id to '';